requirements.txt
torch
torchvision
pillow
transformers
opencv-python-headless
scikit-image
Flask
werkzeug
numpy
requests
openai